requirements.txt
langchain
langchain-community
langchain-core
langchain-classic
langchain-text-splitters
langgraph
ipykernel
langchain-google-genai
pymupdf
arxiv
huggingface-hub
pydantic
qdrant-client 
sentence-transformers
tavily-python 
tiktoken
tokenizers    
rank-bm25  